+++ planning/motion_velocity_planner/autoware_motion_velocity_obstacle_stop_module/src/obstacle_stop_module.hpp
@@ -46,101 +46,6 @@
 
 namespace autoware::motion_velocity_planner
 {
-using autoware::universe_utils::getOrDeclareParameter;
-
-namespace
-{
-template <typename T>
-bool is_in_vector(const T variable, const std::vector<T> & vec)
-{
-  return std::find(vec.begin(), vec.end(), variable) != vec.end();
-}
-
-double calc_minimum_distance_to_stop(
-  const double initial_vel, const double max_acc, const double min_acc)
-{
-  if (initial_vel < 0.0) {
-    return -std::pow(initial_vel, 2) / 2.0 / max_acc;
-  }
-
-  return -std::pow(initial_vel, 2) / 2.0 / min_acc;
-}
-
-autoware::universe_utils::Point2d convert_point(const geometry_msgs::msg::Point & p)
-{
-  return autoware::universe_utils::Point2d{p.x, p.y};
-}
-
-std::vector<TrajectoryPoint> resample_trajectory_points(
-  const std::vector<TrajectoryPoint> & traj_points, const double interval)
-{
-  const auto traj = autoware::motion_utils::convertToTrajectory(traj_points);
-  const auto resampled_traj = autoware::motion_utils::resampleTrajectory(traj, interval);
-  return autoware::motion_utils::convertToTrajectoryPointArray(resampled_traj);
-}
-
-std::vector<PredictedPath> resample_highest_confidence_predicted_paths(
-  const std::vector<PredictedPath> & predicted_paths, const double time_interval,
-  const double time_horizon, const size_t num_paths)
-{
-  std::vector<PredictedPath> sorted_paths = predicted_paths;
-
-  // Sort paths by descending confidence
-  std::sort(
-    sorted_paths.begin(), sorted_paths.end(),
-    [](const PredictedPath & a, const PredictedPath & b) { return a.confidence > b.confidence; });
-
-  std::vector<PredictedPath> selected_paths;
-  size_t path_count = 0;
-
-  // Select paths that meet the confidence thresholds
-  for (const auto & path : sorted_paths) {
-    if (path_count < num_paths) {
-      selected_paths.push_back(path);
-      ++path_count;
-    }
-  }
-
-  // Resample each selected path
-  std::vector<PredictedPath> resampled_paths;
-  for (const auto & path : selected_paths) {
-    if (path.path.size() < 2) {
-      continue;
-    }
-    resampled_paths.push_back(
-      autoware::object_recognition_utils::resamplePredictedPath(path, time_interval, time_horizon));
-  }
-
-  return resampled_paths;
-}
-
-template <class T>
-std::vector<T> concat_vectors(
-  const std::vector<T> & first_vector, const std::vector<T> & second_vector)
-{
-  std::vector<T> concatenated_vector;
-  concatenated_vector.insert(concatenated_vector.end(), first_vector.begin(), first_vector.end());
-  concatenated_vector.insert(concatenated_vector.end(), second_vector.begin(), second_vector.end());
-  return concatenated_vector;
-}
-
-double calc_dist_to_bumper(const bool is_driving_forward, const VehicleInfo & vehicle_info)
-{
-  if (is_driving_forward) {
-    return std::abs(vehicle_info.max_longitudinal_offset_m);
-  }
-  return std::abs(vehicle_info.min_longitudinal_offset_m);
-}
-
-Float64Stamped create_float64_stamped(const rclcpp::Time & now, const float & data)
-{
-  Float64Stamped msg;
-  msg.stamp = now;
-  msg.data = data;
-  return msg;
-}
-}  // namespace
-
 class ObstacleStopModule : public PluginModuleInterface
 {
 public:
